+++ index.js
@@ -1,11 +1,6 @@
-'use strict'
-const cmd = require('./lib/cmd')
-const prepare = require('./lib/prepare')
-const ingest = require('./lib/ingest')
-
-const actions = {
-  prepare: prepare,
-  slurp: ingest
-}
-
-actions[cmd.command](cmd)
+require('yargs')
+  .usage('fatstraw <cmd> [options]')
+  .commandDir('./commands')
+  .help('help')
+  .epilog('Made with ♥️  by Esri DC R&D')
+  .argv